fixes #102: add GenericParser for tracking extra state while parsing
